A bunch of stuff to reduce compile times
But it didn't really do all that much :/

[Finishes #169669112]

+++ Tethys/Feeds/Editing Feeds/FeedDetailView.swift
@@ -24,7 +24,7 @@ public final class FeedDetailView: UIView {
         didSet { self.tagsList.recalculateHeightConstraint() }
     }
 
-    public var maxHeight: Int {
+    public var maxHeight: CGFloat {
         get { return self.tagsList.maxHeight }
         set { self.tagsList.maxHeight = newValue }
     }
